workflow_orchestration/pyproject.toml
[tool.poetry]
name = "workflow-orchestration-module"
version = "0.1.0"
description = "Capstone project Orchestration Module MLOps Zoomcamp"
authors = ["Irem Erturk <irem.erturk@xxx.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = "^1.4.2"
prefect = "^2.1.1"
kaggle = "^1.5.12"
opendatasets = "^0.1.22"
prefect-shell = "^0.1.1"
transformers = "^4.21.2"
spacy = "^3.4.1"
fastai = "^2.7.9"
torch = "^1.12.1"
torchvision = "^0.13.1"
mlflow = "^1.28.0"
python-dotenv = "^0.21.0"
#boto3 = "^1.24"
boto3 = "^1.24.85"
minio = "^7.1.12"

[tool.poetry.dev-dependencies]
greenlet = "^1.1.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
profile = "black"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"